refactor(test-suite): avoid generic upgrade planning split
Keep upgrade target semantics in the existing lifecycle code and limit this PR to explicit rollout contract snapshots.
